VSGI loader for GModule-based applications
Provide a loader, namely 'vsgi', for application written as a GModule.
The syntax is similar to gunicorn and make it possible to pass CLI
arguments to the server.

    vsgi module:symbol -- --port=3003

It loads the symbole from the module, and forward arguments and serve
the application. It does not work for CGI due to the one process per
request constraint.

The implementation arguments must be separated by a '--' to make a clear
distinction from 'vsgi' arguments.

The default server technology is libsoup-2.4 and another implementation
can be specified with the '--server' flag.

The directory where the shared library is stored can be specified with
the '--directory' flag, defaulting on system path.

It should be possible to implement a reloader by monitoring changes on
the shared library in a future release.

Provide an example which use the module initialization to setup required
resources.