vsgi: Provide 'wrorte-status-line' and 'wrote-headers' signals
These two signals makes it possible to adapt headers based on the final
status line value.

For example, a caching middleware could check if the response is
cacheable in 'headers-written' and then tee the response body into the
storage.

For that purpose, 'write_head' is split in two implementation-specific
methods: 'write_status_line' and 'write_headers'. It's all performed in
a 'GLib.Once' critical section and copies are passed to prevent side
effects.