vsgi: Remove '!head_written' contract on 'write_head'
With the 'wrote_status_line' and 'write_headers' signals, there is a way
to capture the moment where these has been sent to the user agent, so we
don't need to put that burden on application code anymore.

Remove checks for '!head_written' since it's up to the latter to
determine that.